fixed PM11_HOLD, added RD21
Fixed connection to PM11_ HOLD which had been connected to PM11_1NC instead of PM11_1NO.  Added RD21, added pull-down to give valid logic signals for two of the diode-OR signals.  Will probably need to add pull-downs to all of the diode-OR logic.  Discovered a bug in the Verilog clutch but did not fix it.